Allow manual setting of current ref.
For the post receive hook, the branch we are currently on is
passed in as stdin.  If we use the 'git branch' method, we will
not be on the correct branch (aamaf, we'll be on a detatched head).
This lets us specify the name of the branch which can be gathered
from the refspec passed in to the hook.